# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "REST API documentation",
    "description": "A voting system for deciding where to have lunch. Course TopJava</a> (Graduation project solution)",
    "contact": {
      "name": "Andrey Kozlov",
      "email": "avgur33@gmail.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "basicAuth": []
    }
  ],
...
  "paths": {
    "/api/root/vote": {
      "put": {
        "tags": [
          "RootController"
        ],
        "summary": "Update vote for authenticated user",
        "description": "if user not voting yet - create vote",
        "operationId": "update",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "query",
            "description": "The id of restaurant.",
            "required": true,
            "content": {
              "*/*": {
                "example": 1
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "post": {
        "tags": [
          "RootController"
        ],
        "summary": "Create vote for authenticated user",
        "operationId": "createVoteWithLocation",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "query",
            "description": "The id of restaurant.",
            "required": true,
            "content": {
              "*/*": {
                "example": 1
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vote"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/root": {
      "get": {
        "tags": [
          "RootController"
        ],
        "summary": "Get menus and restaurants for today vote",
        "operationId": "get",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "description": "Page number ",
            "required": true,
            "content": {
              "*/*": {
                "example": 0
              }
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Page size ",
            "required": true,
            "content": {
              "*/*": {
                "example": 10
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MenuTo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/root/vote/user/history": {
      "get": {
        "tags": [
          "RootController"
        ],
        "summary": "Get users history of voting",
        "operationId": "getAllVotes",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "description": "Start date. Format yyyy-MM-dd.",
            "required": false,
            "content": {
              "*/*": {
                "example": "2020-02-21"
              }
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "End date. Format yyyy-MM-dd.",
            "required": false,
            "content": {
              "*/*": {
                "example": "2022-02-21"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VoteTo"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/root/vote/result/history": {
      "get": {
        "tags": [
          "RootController"
        ],
        "summary": "Get history of voting",
        "operationId": "getResultHistory",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "description": "Start date. Format yyyy-MM-dd.",
            "required": false,
            "content": {
              "*/*": {
                "example": "2020-02-21"
              }
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "End date. Format yyyy-MM-dd.",
            "required": false,
            "content": {
              "*/*": {
                "example": "2022-02-21"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RestaurantTo"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LocalTime": {
        "type": "object",
        "properties": {
          "hour": {
            "type": "integer",
            "format": "int32"
          },
          "minute": {
            "type": "integer",
            "format": "int32"
          },
          "second": {
            "type": "integer",
            "format": "int32"
          },
          "nano": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Vote": {
        "required": [
          "regDate",
          "regTime"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID for new entity must be null",
            "format": "int32"
          },
          "regDate": {
            "type": "string",
            "format": "date"
          },
          "regTime": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "userId": {
            "type": "integer",
            "format": "int32",
            "writeOnly": true
          }
        }
      },
...
      "MenuTo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "dishes": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
      "VoteTo": {
        "required": [
          "regDate",
          "regTime",
          "restaurantLocation",
          "restaurantName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "regDate": {
            "type": "string",
            "format": "date"
          },
          "regTime": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "restaurantName": {
            "maxLength": 100,
            "minLength": 2,
            "type": "string"
          },
          "restaurantLocation": {
            "maxLength": 100,
            "minLength": 2,
            "type": "string"
          }
        }
      },
      "RestaurantTo": {
        "required": [
          "location",
          "name",
          "votes"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 100,
            "minLength": 2,
            "type": "string"
          },
          "location": {
            "maxLength": 100,
            "minLength": 5,
            "type": "string"
          },
          "votes": {
            "type": "integer",
            "format": "int32"
          },
          "votesHistory": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}